Introducing cernan::thread.
A new abstraction serving mio enbabled threads.
